\name{protoVS-package}
\alias{protoVS-package}
\alias{protoVS}
\docType{package}
\title{\packageTitle{protoVS}}
\description{\packageDescription{protoVS}}
\section{Package Content}{\packageIndices{protoVS}}
\author{\packageAuthor{protoVS}}
\section{Maintainer}{\packageMaintainer{protoVS}}
\keyword{package}
