% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewshed_comparison.R
\name{viewshed_comparison}
\alias{viewshed_comparison}
\title{Viewshed Benchmark}
\usage{
viewshed_comparison(
  observer,
  dsm_rast,
  dtm_rast,
  max_distance_vec = 1:5,
  observer_height = 1.7,
  sample_size = 10,
  ncores = 1,
  progress = FALSE
)
}
\arguments{
\item{observer}{object of class \code{sf}; Observer locations from which viewsheds will be calculated}

\item{dsm_rast}{object of class \code{\link[terra]{rast}}; \code{\link[terra]{rast}} of the DSM}

\item{dtm_rast}{object of class \code{\link[terra]{rast}}; \code{\link[terra]{rast}} of the DTM}

\item{max_distance_vec}{numeric vector; Buffer distances to calculate the viewsheds}

\item{observer_height}{numeric > 0; Height of the observer (e.g. 1.7 meters)}

\item{sample_size}{numeric > 0; Number of times to evaluate the viewshed calculation at each distance level.}

\item{ncores}{integer; he number of cores to use, i.e. at most how many child processes will be run simultaneously}

\item{progress}{logical; Show progress bar and computation time?}
}
\value{
object of class \code{\link[tibble]{tibble}};
}
\description{
Compares computation times of the traditional against our novel viewshed algorithm.
At each distance level a binary viewshed for each observer point will be calculated based on a Digital Surface Model (DSM).
}
